changes to get code running
Workarounds to get this code to compile and run successfully on Windows 11 using Visual Studio 2022 and CUDA 11.7. No guarantees about performance as a lot has changed since the RT kernels were written, and there are visible artifacts in the results. speculative_while_while also is definitely doing something _wrong_ and has visible ghosting from rays at a different perspective in the resulting image. But hey, it runs, and that's more than I could get it to do before.